+++ abci.go
@@ -111,9 +111,12 @@ func (c CheckResult) ToABCI() abci.ResponseCheckTx {
 	}
 }
 
-// TickResult allows the Ticker to modify the validator set
+// TickResult contains the result of the tasks execution.
 type TickResult struct {
-	Diff []ValidatorUpdate
+	// List of task IDs that were executed. Task is considered executed
+	// when processing it caused any change to the state (even if it is
+	// only removing the task from the queue and no other change).
+	Executed [][]byte
 }
 
 // DeliverTxError converts any error into a abci.ResponseDeliverTx, preserving