+++ handler.go
@@ -34,12 +34,6 @@ type Decorator interface {
 	Deliver(ctx Context, store KVStore, tx Tx, next Deliverer) (*DeliverResult, error)
 }
 
-// Ticker is a method that is called the beginning of every block,
-// which can be used to perform periodic or delayed tasks
-type Ticker interface {
-	Tick(ctx Context, store KVStore) (TickResult, error)
-}
-
 // Registry is an interface to register your handler,
 // the setup side of a Router
 type Registry interface {