+++ CHANGELOG.md
@@ -2,16 +2,24 @@
 
 ## HEAD
 
-- Add `owner` index to bnsd `x/username` to be able to query tokens by owner.
-- Allow empty targets in bnsd `x/username` to enable name reservation.
+- A new extension `x/cron` is added. It allows to configure weave application
+  to be able to schedule messages for future execution.
+- Proposals created with `x/gov` are having their tally executed automatically
+  after the voting time is over. This is possible thanks to `x/cron` extension.
+  This is not a breaking change because manual tally, although pointless, is
+  still possible.
 
 Breaking changes
 
+- `weave.Ticker` interface was updated.
+- Add `owner` index to bnsd `x/username` to be able to query tokens by owner.
+- Allow empty targets in bnsd `x/username` to enable name reservation.
 - Username address type in `x/username` extension was changed from `[]byte` to
   `string`. Instead of base64 encoded value, a valid string is stored as the
   address.
 - Some of the query paths in the `x/gov` package were updated to follow the
   naming convention.
+
 
 
 ## 0.18.0